pyproject.toml
[project]
name = "moods"
version = "0.1.0"
description = "Shared mood-tracking app"
requires-python = ">=3.12"
dependencies = [
    "starlette",
    "ariadne",
    "asyncpg",
    "aiosql",
    "dynaconf",
    "yoyo-migrations",
    "uvicorn[standard]",
    "aiodataloader>=0.4.3",
    "psycopg2-binary>=2.9.11",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/moods"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
